Added messages to BusStopPage
It is now possible to read messages set by Tfl.
Also imroved a little on the looks of BusStopPage
and JourneyProgressPage. Had to modify ArrivalsModel
because when model reseted itself it also reseted the
view which in turn restarted the animation of the message.
It was also neccessary to call reset from qml
(StopHeader and RunningText) in order to repaint
the header.