Ranking added to stopstable
All favorite stops have rank which determines their order they are
presented to user. All non favorite stops are with no rank and placed on
top of all other, as they are more relevant than favorite stops. New
favorite stops are placed on the top of the list under non favorites,
but user will be able to move them further down if so they choose. When
user unfavorites a stop it's rank is removed and the gap is filled by
other stops.