Fixed Drag and Drop on DeparturePage.qml
User can now only drag an item if it's a favorite and user can only drop
an item in front of a favorite item. This is to simplify the problem
with unfavorite items, they NEED to be on the top for usability reasons
and therefore there is no point in allowing user to move them if they
will be restored once the page is opened again.